pyproject.toml
[project]
name = "streamlitapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "japanize-matplotlib>=1.1.3",
    "jupyterlab>=4.4.1",
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "seaborn>=0.13.2",
    "streamlit>=1.44.1",
]
